refactor: Move canister manager types in their own module
